Working version is done.

+++ configure.d/i3blocks.sh
@@ -4,7 +4,9 @@ arg="$@"
 utils="/usr/lib/`basename $(realpath $0)`"
 
 # Link utils.
-for i in $utils/* ; do
-	rm -f "$i"
-	ln -s $arg "$i" $XDG_CONFIG_HOME/i3blocks
-done
+if [ -d "$utils" ] ; then
+	for i in $utils/* ; do
+		rm -f "$i"
+		ln -s $arg "$i" $XDG_CONFIG_HOME/i3blocks
+	done
+fi
